upgrade dependencies
