+++ project/build.properties
@@ -1,2 +1,2 @@
 # suppress inspection "UnusedProperty"
-sbt.version = 1.9.7
+sbt.version = 1.9.8
