+++ project/builds.sbt
@@ -1,3 +1,3 @@
 libraryDependencies ++= List(
-  "com.lihaoyi" %% "upickle" % "3.1.3"
+  "com.lihaoyi" %% "upickle" % "3.1.4"
 )
