+++ project/sbt-avro.sbt
@@ -1,3 +1,3 @@
-addSbtPlugin("com.github.sbt" % "sbt-avro" % "3.4.3")
+addSbtPlugin("com.github.sbt" % "sbt-avro" % "3.4.4")
 
 libraryDependencies += "org.apache.avro" % "avro-compiler" % "1.11.3"
